test(pdf): merge empty-results tests into pdf.test.ts
The carved-off pdf-empty-results.test.ts implied a module that doesn't
exist. Test files mirror the module under test, so the empty-results
tests belong inside pdf.test.ts.

Mock pdf-parse with a jest.requireActual passthrough so the existing
real-fixture tests keep running unmodified, while the new describe
block overrides per-call with mockResolvedValueOnce.
