+++ test/playbook/sources/pdf.test.ts
@@ -1,9 +1,25 @@
+// Wrap pdf-parse in a passthrough jest.fn so the rest of the suite keeps
+// running real fixtures while the empty-results describe block can override
+// individual calls with mockResolvedValueOnce.
+jest.mock("pdf-parse", () => {
+  const actual = jest.requireActual("pdf-parse") as (
+    buffer: Buffer,
+  ) => Promise<{ text: string; numpages: number }>;
+  return jest.fn(actual);
+});
+
+import { writeFileSync, unlinkSync } from "fs";
+import { tmpdir } from "os";
+import path from "path";
+
+import { glob } from "glob";
+import pdfParse from "pdf-parse";
+
 import { PDFPlaybookSource } from "../../../src/playbook/sources/pdf";
 import { PdfParseError, PDF_PARSE_STAGES } from "../../../src/playbook/sources/errors";
 import { root } from "../../../src/logging";
-import path from "path";
-import { glob } from "glob";
-import { tmpdir } from "os";
+
+const mockPdfParse = pdfParse as unknown as jest.Mock;
 
 describe("PDFPlaybookSource - PlaybookSource interface implementation", () => {
   const fixturesDir = path.join(__dirname, "../../fixtures");
@@ -182,6 +198,70 @@ describe("PDFPlaybookSource - PlaybookSource interface implementation", () => {
           const stage = (error as PdfParseError).stage;
           expect(PDF_PARSE_STAGES).toContain(stage);
         }
+      });
+    });
+  });
+
+  describe("empty extraction results", () => {
+    const stubFixturePath = path.join(tmpdir(), "pdf-empty-results.test.pdf");
+
+    const stubResult = (text: string) => ({
+      text,
+      numpages: 1,
+      numrender: 1,
+      info: {},
+      metadata: null,
+      version: "v1.10.100",
+    });
+
+    beforeAll(() => {
+      // pdf-parse is mocked per-test via mockResolvedValueOnce, so the file's
+      // contents are never extracted. readFileSync still runs in load(), so
+      // the file must exist on disk.
+      writeFileSync(stubFixturePath, "stub");
+    });
+
+    afterAll(() => {
+      try {
+        unlinkSync(stubFixturePath);
+      } catch {
+        // ignore — best-effort cleanup
+      }
+    });
+
+    it("throws PdfParseError tagged 'split' when extracted text yields no playbook sections", async () => {
+      mockPdfParse.mockResolvedValueOnce(stubResult("tiny"));
+
+      const source = new PDFPlaybookSource(stubFixturePath, "pdf");
+      const promise = source.load();
+
+      await expect(promise).rejects.toBeInstanceOf(PdfParseError);
+      await expect(promise).rejects.toMatchObject({
+        stage: "split",
+        message: expect.stringContaining("no playbook sections"),
+      });
+    });
+
+    it("throws PdfParseError tagged 'section-parse' when no section produces a valid playbook", async () => {
+      // Long enough to clear minSectionLength (100), but lacks every archetype
+      // pattern and every playbook indicator ("Choose Your Nature", "Starting
+      // Moves", or the Background/vagabond pair) — every section is rejected
+      // as non-playbook.
+      const noisyText = (
+        "this is a long lorem ipsum style passage with no proper noun patterns " +
+        "and no special markers and no caps lines so the parser will not find an " +
+        "archetype here despite the text being long enough to count as a section "
+      ).repeat(2);
+
+      mockPdfParse.mockResolvedValueOnce(stubResult(noisyText));
+
+      const source = new PDFPlaybookSource(stubFixturePath, "pdf");
+      const promise = source.load();
+
+      await expect(promise).rejects.toBeInstanceOf(PdfParseError);
+      await expect(promise).rejects.toMatchObject({
+        stage: "section-parse",
+        message: expect.stringContaining("no valid playbooks"),
       });
     });
   });
